fix(telos): keep payload classifier within line budget [impact-checked]
Extracts typed-action payload harm classification into a small telos-owned helper module re-exported by telos_gates, keeping telos_gates under Rule 10 while preserving the drift fix from the prior commit.

Verification: ruff check dharma_swarm/ontology.py dharma_swarm/telos_gates.py dharma_swarm/telos_payload_classifier.py tests/test_ontology_telos_hardwire.py tests/test_logic_layer.py; pytest -q tests/test_ontology_telos_hardwire.py tests/test_ontology_registry.py tests/test_logic_layer.py::TestApplyAction tests/test_api.py::TestOntologyActions -q; python3 scripts/governance/check_module_budget.py --base-ref main --head-ref HEAD; make test-contracts
